pyproject.toml
[project]
name = "blog-radar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "click>=8.3.1",
    "feedparser>=6.0.12",
    "openai>=2.9.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]
